+++ src/Sentry/Platforms/Android/SentrySdk.cs
@@ -94,11 +94,8 @@ public static partial class SentrySdk
                     o.CacheDirPath = Path.Combine(cacheDirectoryPath, "android");
                 }
 
-                var javaTags = o.Tags;
-                foreach (var tag in options.DefaultTags)
-                {
-                    javaTags.Add(tag);
-                }
+                // NOTE: Tags in options.DefaultTags should not be passed down, because we already call SetTag on each
+                //       one when sending events, which is relayed through the scope observer.
 
                 if (options.HttpProxy is System.Net.WebProxy proxy)
                 {
@@ -160,8 +157,8 @@ public static partial class SentrySdk
                 o.ReadTimeoutMillis = (int)options.Android.ReadTimeout.TotalMilliseconds;
 
                 // In-App Excludes and Includes to be passed to the Android SDK
-                options.Android.InAppExclude?.ToList().ForEach(x => o.AddInAppExclude(x));
-                options.Android.InAppInclude?.ToList().ForEach(x => o.AddInAppInclude(x));
+                options.Android.InAppExcludes?.ForEach(x => o.AddInAppExclude(x));
+                options.Android.InAppIncludes?.ForEach(x => o.AddInAppInclude(x));
 
                 // These options are intentionally set and not exposed for modification
                 o.EnableExternalConfiguration = false;