+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Enable Scope Sync for iOS ([#1834](https://github.com/getsentry/sentry-dotnet/pull/1834))
 - Add API for deliberately crashing an app ([#1842](https://github.com/getsentry/sentry-dotnet/pull/1842))
 - Add Mac Catalyst target ([#1848](https://github.com/getsentry/sentry-dotnet/pull/1848))
+- Add and configure options for the iOS SDK ([#1849](https://github.com/getsentry/sentry-dotnet/pull/1849))
 
 ### Fixes